requirements.txt
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.1
clickclick==20.10.2
cloudevents==1.2.0
colorama==0.4.4
connexion==2.7.0
deprecation==2.1.0
Flask==2.0.2
greenlet==1.1.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.8.1
importlib-resources==5.4.0
inflection==0.5.1
isodate==0.6.0
itsdangerous==2.0.1
Jinja2==3.0.1
jsonschema==4.2.0
MarkupSafe==2.0.1
openapi-schema-validator==0.1.5
openapi-spec-validator==0.3.1
packaging==20.9
PyMySQL==1.0.2
pyparsing==2.4.7
pyrsistent==0.18.0
python-dotenv==0.19.1
PyYAML==6.0
redis==3.5.3
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.18
typing-extensions==3.10.0.2
urllib3==1.26.7
watchdog==1.0.2
Werkzeug==2.0.1
zipp==3.6.0
